INTLY-106 Add list polling on the frontend
Currently, when the user goes to the application in walkthroughs 1
and 1A they are made to manually refresh to see updates that are
coming from Fuse. It would be nice if a user did not need to
manually refresh.

This change adds polling to the fruit list endpoint to remove the
need for the user to manually refresh.

Verification:
- Deploy the CRUD app to OpenShift using the steps described in the
README
- Open the CRUD app in two tabs or windows
- Ensure an item created in one tab is shown in the other without
needing a refresh
- Ensure create/edit/delete still work as expected