Revert "fix swagger spec"
This reverts commit 3fbf65bfda45af0aa348c66aede720cf637929bb.

The referenced commit made changes to the Swagger spec of the app
that resulted in its import into fuse no longer working. However,
in the installer, even though the image tag references a commit
after this commit, none of these changes are included.

My guess is that a manual build of the image was done using an old
commit but given a later commit tag. So the changes in the reverted
commit were never included in Integreatly.